DATABASE/API/requirements.txt
# api/requirements.txt

Flask
psycopg2-binary
bcrypt  